Move models in to models/ folder
