+++ include/fullscore/models/note_playback_info.h
