+++ include/fullscore/transforms/insert_note_transform.h
@@ -6,7 +6,7 @@
 
 #include <fullscore/transform_base.h>
 
-#include <fullscore/note.h>
+#include <fullscore/models/note.h>
