+++ src/measure.cpp
@@ -1,8 +1,8 @@
 
 // empty
 
-#include <fullscore/measure.h>
-#include <fullscore/note.h>
+#include <fullscore/models/measure.h>
+#include <fullscore/models/note.h>
 #include <allegro_flare/useful.h>
