+++ src/music_engraver.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/note_playback_info.h>
+#include <fullscore/models/note_playback_info.h>
 
 
 
@@ -11,7 +11,7 @@
 #include <allegro_flare/useful.h>
 #include <allegro_flare/drawing_interfaces/drawing_interface_allegro5.h>
 
-#include <fullscore/note.h>
+#include <fullscore/models/note.h>
 #include <fullscore/music_engraver.h>
