+++ src/note.cpp
@@ -2,7 +2,7 @@
 
 
 #include <sstream>
-#include <fullscore/note.h>
+#include <fullscore/models/note.h>
